Added a filter for Archive files in method mostNestedEquivalent
If this filter is not present, search results of the FileSearch including archive Files can not be opened correctly in some cases. If another archive file that is more nested is present, the search result will lead to the more nested archive file, resulting in the wrong file to be opened. With the filter the correct resource is returned in those cases.