adjust MessageDialog for deleting zip files
Deleting an opened zip file also deletes the resource in the local file system. An opened zip file is represented as a linked folder which shows a specific Messagedialog when being deleted indicating that only the link will be deleted. In this case the delete operation of zip files also deletes the resource in the local file system.

Please mind that one dialog exists in JDT which should also be adjusted similarily.